requirements.txt
python-nmap
pyPdf
pygeoip
mechanize
BeautifulSoup4
docopts
colorama
